Moved database_objects to their own files for visual clarity
