+++ common/src/index.ts
@@ -1,68 +1,83 @@
 import {
-  ApiKey,
-  ApiKeySchema,
-  Blog,
-  BlogSchema,
-  CreateAccount,
   CreateAccountSchema,
-  CreateBlog,
-  CreateBlogSchema,
-  CreateEvent,
-  CreateEventSchema,
-  CreateLocation,
-  CreateLocationSchema,
-  CreatePrice,
-  CreatePriceSchema,
-  CreateProduction,
-  CreateProductionSchema,
-  CreateTag,
-  CreateTagSchema,
-  Event,
-  EventSchema,
-  FilterEvent,
-  FilterEventSchema,
-  FilterProduction,
-  FilterProductionSchema,
-  Location,
-  LocationSchema,
-  Price,
-  PriceSchema,
-  Production,
-  ProductionSchema,
-  PublicAccount,
   PublicAccountSchema,
-  Tag,
-  TagSchema,
-  UpdateAccount,
   UpdateAccountSchema,
-  UpdateBlog,
-  UpdateBlogSchema,
-  UpdateEvent,
-  UpdateEventSchema,
-  UpdateLocation,
-  UpdateLocationSchema,
-  UpdatePrice,
-  UpdatePriceSchema,
-  UpdateProduction,
-  UpdateProductionSchema,
-  UpdateTag,
-  UpdateTagSchema,
-  VerifyApiKey,
+  UpdateAccount,
+  PublicAccount,
+  CreateAccount,
+} from "./objects/accounts";
+import {
+  ApiKeySchema,
   VerifyApiKeySchema,
-  PaginationFilterSchema,
-  PaginationFilter,
-  Language,
+  ApiKey,
+  VerifyApiKey,
+} from "./objects/api.keys";
+import {
+  BlogSchema,
+  BlogViewSchema,
+  UpdateBlogSchema,
+  CreateBlogSchema,
+  Blog,
+  BlogView,
+  CreateBlog,
+  UpdateBlog,
+} from "./objects/blogs";
+import {
+  EventSchema,
+  CreateEventSchema,
+  UpdateEventSchema,
+  FilterEventSchema,
+  Event,
+  CreateEvent,
+  UpdateEvent,
+  FilterEvent,
+} from "./objects/events";
+import { PaginationFilterSchema, PaginationFilter } from "./objects/extra";
+import {
   LanguageQuerySchema,
   DEFAULT_LANGUAGE,
   SUPPORTED_LANGUAGES,
+  Language,
   LanguageQuery,
-  ProductionView,
+} from "./objects/language";
+import {
+  LocationSchema,
+  CreateLocationSchema,
+  UpdateLocationSchema,
+  Location,
+  CreateLocation,
+  UpdateLocation,
+} from "./objects/locations";
+import {
+  PriceSchema,
+  CreatePriceSchema,
+  UpdatePriceSchema,
+  Price,
+  CreatePrice,
+  UpdatePrice,
+} from "./objects/prices";
+import {
+  ProductionSchema,
   ProductionViewSchema,
-  TagView,
+  CreateProductionSchema,
+  UpdateProductionSchema,
+  FilterProductionSchema,
+  Production,
+  ProductionView,
+  CreateProduction,
+  UpdateProduction,
+  FilterProduction,
+} from "./objects/productions";
+import {
+  TagSchema,
   TagViewSchema,
-  BlogView,
-  BlogViewSchema,
-} from "./database_objects";
+  CreateTagSchema,
+  UpdateTagSchema,
+  Tag,
+  TagView,
+  CreateTag,
+  UpdateTag,
+} from "./objects/tags";
 
 // list of all exports: (this way only need to import this file.)
 export {
