+++ common/src/objects/tags.ts
@@ -1,0 +1,25 @@
+import { z } from "zod";
+import { LocalizedStringSchema } from "./language";
+
+export const TagSchema = z.object({
+  id: z.number(),
+  tag: LocalizedStringSchema,
+  created_at: z.iso.datetime(),
+  updated_at: z.iso.datetime(),
+  legacy_id: z.string().nullable(),
+});
+export const TagViewSchema = TagSchema.extend({
+  tag: z.string(),
+});
+
+export const CreateTagSchema = TagSchema.omit({
+  id: true,
+  created_at: true,
+  updated_at: true,
+});
+export const UpdateTagSchema = TagSchema.partial();
+
+export type Tag = z.infer<typeof TagSchema>;
+export type TagView = z.infer<typeof TagViewSchema>;
+export type CreateTag = z.infer<typeof CreateTagSchema>;
+export type UpdateTag = z.infer<typeof UpdateTagSchema>;
